renamed to similarity
