+++ pymove_osmnx/__init__.py
@@ -4,7 +4,8 @@ PyMove-OSMnx
 Provides processing, map matching and visualization of trajectories and other
 spatial-temporal data
 """
+from ._version import __version__
 
 from .utils import interpolate
-from .utils import LongestCommonSubSequence
+from .utils import similarity
 from .core import map_matching_osmnx
